+++ scripts/rollHandlers/sfrpg/sfrpg-base.js
@@ -1,5 +1,5 @@
-import { RollHandler } from "../rollHandler.js"
-import * as settings from "../../settings.js";
+import { RollHandler } from '../rollHandler.js'
+import * as settings from '../../settings.js';
 
 export class RollHandlerBaseSfrpg extends RollHandler {
     constructor() {
@@ -7,7 +7,7 @@ export class RollHandlerBaseSfrpg extends RollHandler {
     }
 
     /** @override */
-    doHandleActionEvent(event, encodedValue) {
+    async doHandleActionEvent(event, encodedValue) {
         let payload = encodedValue.split('|');
         
         if (payload.length != 3) {
@@ -19,29 +19,35 @@ export class RollHandlerBaseSfrpg extends RollHandler {
         let actionId = payload[2];
 
         switch (macroType) {
-            case "ability":
+            case 'ability':
                 this.rollAbilityMacro(event, tokenId, actionId);
                 break;
-            case "skill":
+            case 'skill':
                 this.rollSkillMacro(event, tokenId, actionId);
                 break;
-            case "save":
+            case 'save':
                 this.rollSaveMacro(event, tokenId, actionId);
                 break;
-            case "abilitySave":
+            case 'abilitySave':
                 this.rollAbilitySaveMacro(event, tokenId, actionId);
                 break;
-            case "abilityCheck":
+            case 'abilityCheck':
                 this.rollAbilityCheckMacro(event, tokenId, actionId);
                 break;
-            case "item":
-            case "spell":
-            case "feat": 
+            case 'item':
+            case 'spell':
+            case 'feat': 
+            case 'starshipWeapon':
                 if (this.isRenderItem())
                     this.doRenderItem(tokenId, actionId);
                 else
                     this.rollItemMacro(event, tokenId, actionId);
                 break;
+            case 'shields':
+                this._handleShields(event, tokenId, actionId);
+                break;
+            case 'crewAction':
+                this._handleCrewAction(event, tokenId, actionId);
             default:
                 break;
         }
@@ -76,7 +82,7 @@ export class RollHandlerBaseSfrpg extends RollHandler {
             return;
         }
         
-        if (item.data.type === "spell")
+        if (item.data.type === 'spell')
             return actor.useSpell(item);
             
         return item.roll();
@@ -84,5 +90,53 @@ export class RollHandlerBaseSfrpg extends RollHandler {
 
     needsRecharge(item) {
         return (item.data.data.recharge && !item.data.data.recharge.charged && item.data.data.recharge.value);
+    }
+
+    async _handleShields(event, tokenId, actionId) {
+        const actor = super.getActor(tokenId);
+        let payload = actionId.split('.');
+
+        const side = payload[0];
+        let shieldChange = parseInt(payload[1]);
+        if (shieldChange === NaN)
+            return;
+
+        const shields = actor.data.data.attributes.shields;
+        const shield = shields[side];
+
+        let newValue;
+        if (shieldChange < 0) {
+            newValue = Math.clamped(shield.value + shieldChange, 0, shield.max);
+        } else {
+            newValue = this._calcPossibleIncrease(shields, shield, shieldChange);
+        }
+
+        if (newValue === shield.value)
+            return;
+
+        const update = {data: {attributes: {shields: {}}}};
+        update.data.attributes.shields[side] = {value: newValue};
+
+        await actor.update(update);
+    }
+    
+    _calcPossibleIncrease(shields, shield, change) {
+        const overallPossible = shields.max - shields.value >= 0 ? shields.max - shields.value : 0;
+        const localPossible = shield.max - shield.value >= 0 ? shield.max - shield.value : 0;
+
+        let possibleChange = change;
+
+        if (change > overallPossible)
+            possibleChange = overallPossible;
+
+        if (change > localPossible)
+            possibleChange = localPossible;
+
+        return shield.value + possibleChange;
+    }
+
+    _handleCrewAction(event, tokenId, actionId) {
+        const actor = super.getActor(tokenId);
+        actor.useStarshipAction(actionId);
     }
 }