+++ scripts/rollHandlers/swade/swade-base.js
@@ -33,6 +33,9 @@ export class RollHandlerBaseSwade extends RollHandler {
             case 'benny':
                 this._adjustBennies(event, actor, actionId);
                 break;
+            case 'gmBenny':
+                await this._adjustGmBennies(event, actor, actionId);
+                break;
             case 'attribute':
                 this._rollAttribute(event, actor, actionId);
                 break;
@@ -65,11 +68,75 @@ export class RollHandlerBaseSwade extends RollHandler {
 
     /** @private */
     _adjustBennies(event, actor, actionId) {
-        if (actionId === 'spend')
+        if (actionId === 'spend') {
             actor.spendBenny();
+            this._showDiceBenny();
+        }
 
         if (actionId === 'get')
             actor.getBenny();
+    }
+
+    /** @private */
+    async _adjustGmBennies(event, actor, actionId) {
+        let user = game.user;
+        if(!user.isGM)
+            return;
+
+        const benniesValue = user.getFlag('swade', 'bennies');
+        if (actionId === 'spend') {
+            if (benniesValue == 0)
+                return;
+
+            if (game.settings.get('swade', 'notifyBennies')) {
+                await this._createGmSpendMessage(user);
+                this._showDiceBenny();
+            }
+
+            await user.setFlag('swade', 'bennies', benniesValue - 1);
+        }
+
+        if (actionId === 'get') {
+            await user.setFlag('swade', 'bennies', benniesValue+1);
+            if (game.settings.get('swade', 'notifyBennies')) {
+               await this._createGmGetMessage(user);
+            }
+            ui['players'].render(true);
+        }
+
+        Hooks.callAll('forceUpdateTokenActionHUD');
+    }
+
+    async _createGmSpendMessage(user) {
+        let message = await renderTemplate(CONFIG.SWADE.bennies.templates.spend, {
+            target: user,
+            speaker: user,
+        });
+
+        let chatData = {
+            content: message,
+        };
+        ChatMessage.create(chatData);
+    }
+
+    async _createGmGetMessage(user) {
+        let message = await renderTemplate(CONFIG.SWADE.bennies.templates.gmadd, {
+            target: user,
+            speaker: user,
+        });
+
+        let chatData = {
+            content: message,
+        };
+        
+        ChatMessage.create(chatData);
+    }
+
+    _showDiceBenny() {
+        if (game.dice3d) {
+            const benny = new Roll('1dB').roll();
+            game.dice3d.showForRoll(benny, game.user, true, null, false);
+        }
     }
 
     /** @private */